requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.2
rank-bm25==0.2.2
sentence-transformers>=2.3.0
faiss-cpu>=1.8.0
streamlit==1.28.2
pytest==7.4.3
pandas==2.1.3
requests==2.31.0
numpy==1.26.2
tqdm==4.66.1
httpx>=0.23.0
